requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
